feat(admin): add authorization checks for destroy actions across multiple resources
- Implemented authorization checks for the destroy action in various ActiveAdmin resources, ensuring that only authorized users can delete records.
- Updated the destroy methods in Attendance, ContactMessage, Event, Membership, OptionType, OptionValue, OrderItem, Order, OrganizerApplication, Partner, Payment, ProductCategory, ProductVariant, Product, Role, RollerStock, Route, User, VariantOptionValue, and Event::Initiation to include authorization logic.
- Enhanced user feedback with success and error messages upon deletion attempts, improving the overall user experience.

+++ app/admin/attendances.rb
@@ -97,6 +97,7 @@ ActiveAdmin.register Attendance do
   controller do
     def destroy
       @attendance = resource
+      authorize @attendance, :destroy?
       if @attendance.destroy
         redirect_to collection_path, notice: "La participation ##{@attendance.id} a été supprimée avec succès."
       else
